+++ jhdf/src/main/java/io/jhdf/dataset/DatasetBase.java
@@ -3,7 +3,7 @@
  *
  * http://jhdf.io
  *
- * Copyright (c) 2022 James Mudd
+ * Copyright (c) 2023 James Mudd
  *
  * MIT License see 'LICENSE' file
  */
@@ -27,6 +27,8 @@ import io.jhdf.object.message.DataSpaceMessage;
 import io.jhdf.object.message.DataTypeMessage;
 import io.jhdf.object.message.FillValueMessage;
 import io.jhdf.storage.HdfBackingStorage;
+import org.apache.commons.lang3.ArrayUtils;
+import org.apache.commons.lang3.time.StopWatch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,6 +37,7 @@ import java.nio.ByteOrder;
 import java.util.Arrays;
 
 import static java.nio.ByteOrder.LITTLE_ENDIAN;
+import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.apache.commons.lang3.ClassUtils.primitiveToWrapper;
 
 public abstract class DatasetBase extends AbstractNode implements Dataset {
@@ -119,16 +122,44 @@ public abstract class DatasetBase extends AbstractNode implements Dataset {
 
 	@Override
 	public Object getData() {
-		logger.debug("Getting data for '{}'...", getPath());
+		logger.info("Getting data for '{}'...", getPath());
 
 		if (isEmpty()) {
 			return null;
 		}
 
+		final StopWatch stopWatch = StopWatch.createStarted();
+
 		final ByteBuffer bb = getDataBuffer();
 		final DataType type = getDataType();
 
-		return DatasetReader.readDataset(type, bb, getDimensions(), hdfBackingStorage);
+		final Object data = DatasetReader.readDataset(type, bb, getDimensions(), hdfBackingStorage);
+		stopWatch.stop();
+
+		logger.info("Finished getting data for [{}] took [{}ms]", getPath(), stopWatch.getTime(MILLISECONDS));
+		return data;
+	}
+
+	@Override
+	public Object getDataFlat() {
+		logger.info("Getting flat data for [{}]...", getPath());
+
+		if (isEmpty()) {
+			return ArrayUtils.EMPTY_OBJECT_ARRAY;
+		}
+
+		final StopWatch stopWatch = StopWatch.createStarted();
+
+		final ByteBuffer bb = getDataBuffer();
+		final DataType type = getDataType();
+
+		int elements = Math.toIntExact(getSize());
+
+		final Object data = DatasetReader.readDataset(type, bb, elements, hdfBackingStorage);
+		stopWatch.stop();
+
+		logger.info("Finished getting flat data for [{}] took [{}ms]", getPath(), stopWatch.getTime(MILLISECONDS));
+		return data;
 	}
 
 	@Override