+++ jhdf/src/test/java/io/jhdf/dataset/CompactDatasetTest.java
@@ -3,14 +3,14 @@
  *
  * http://jhdf.io
  *
- * Copyright (c) 2022 James Mudd
+ * Copyright (c) 2023 James Mudd
  *
  * MIT License see 'LICENSE' file
  */
 package io.jhdf.dataset;
 
 import io.jhdf.HdfFile;
-import io.jhdf.TestUtils;
+import io.jhdf.Utils;
 import io.jhdf.api.Dataset;
 import io.jhdf.exceptions.UnsupportedHdfException;
 import io.jhdf.filter.PipelineFilterWithData;
@@ -87,7 +87,7 @@ class CompactDatasetTest {
 			assertThat(filters, is(empty()));
 
 			Object data = dataset.getData();
-			assertThat(TestUtils.getDimensions(data), is(equalTo(new int[]{10})));
+			assertThat(Utils.getDimensions(data), is(equalTo(new int[]{10})));
 			Object[] flatData = flatten(data);
 			for (int i = 0; i < flatData.length; i++) {
 				// Do element comparison as there are all different primitive numeric types