+++ jhdf/src/test/java/io/jhdf/examples/TestAllFilesBase.java
@@ -3,7 +3,7 @@
  *
  * http://jhdf.io
  *
- * Copyright (c) 2022 James Mudd
+ * Copyright (c) 2023 James Mudd
  *
  * MIT License see 'LICENSE' file
  */
@@ -30,6 +30,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.PathMatcher;
 import java.nio.file.Paths;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -37,7 +38,8 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static io.jhdf.TestUtils.getDimensions;
+import static io.jhdf.TestUtils.flatten;
+import static io.jhdf.Utils.getDimensions;
 import static org.apache.commons.lang3.ArrayUtils.toObject;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.anEmptyMap;
@@ -152,40 +154,60 @@ abstract class TestAllFilesBase {
 		dataset.getData();
 		final Object data = dataset.getData();
 
+		final Object dataFlat = dataset.getDataFlat();
+
 		if (dataset.isEmpty()) {
 			assertThat(data, is(nullValue()));
 			// Empty so should have 0 size
 			assertThat(dataset.getStorageInBytes(), is(equalTo(0L)));
 			// Could still have non-zero size in case no storage is allocated
 			assertThat(dataset.getStorageInBytes(), is(greaterThanOrEqualTo(0L)));
+			assertThat((Object[]) dataFlat, is(emptyArray()));
+			assertThat(getDimensions(dataFlat), is(new int[]{0}));
 		} else if (dataset.isScalar()) {
 			assertThat(toObject(dataset.getDimensions()), is(emptyArray()));
 			assertThat(data.getClass(), is(equalTo(dataset.getJavaType())));
 			// Should have some size
 			assertThat(dataset.getSizeInBytes(), is(greaterThan(0L)));
 			assertThat(dataset.getSize(), is(equalTo(1L)));
+			assertThat(getDimensions(dataFlat), is(new int[]{1}));
+			assertThat(Array.get(dataFlat, 0), is(data));
 		} else if (dataset.isCompound()) {
 			// Compound datasets are currently returned as maps, maybe a custom CompoundDataset might be better in the future...
 			assertThat(data, is(instanceOf(Map.class)));
-			assertThat((Map<String, Object>) data, is(not(anEmptyMap())));
+			Map<String, Object> mapData = (Map<String, Object>) data;
+			Map<String, Object> flatMapData = (Map<String, Object>) dataFlat;
+			assertThat(mapData, is(not(anEmptyMap())));
 			assertThat(dataset.getSizeInBytes(), is(greaterThan(0L)));
 			assertThat(dataset.getSize(), is(greaterThan(0L)));
+			for (String key : mapData.keySet()) {
+				// Don't check nested compound
+				if(mapData.get(key).getClass().isArray()) {
+					assertThat(flatten(mapData.get(key)), is(equalTo(flatten(flatMapData.get(key)))));
+				}
+			}
+
 		} else if (dataset.isVariableLength()) {
 			assertThat(getDimensions(data)[0], is(equalTo(dims[0])));
 			assertThat(dataset.getSizeInBytes(), is(greaterThan(0L)));
 			assertThat(dataset.getSize(), is(greaterThan(0L)));
+			assertThat(getDimensions(dataFlat)[0], is(Arrays.stream(dataset.getDimensions()).reduce(1, Math::multiplyExact)));
 		} else if (dataset.getJavaType().isArray()) {
 			// e.g. Opaque dataset
 			assertThat(getType(data), is(equalTo(dataset.getJavaType().getComponentType())));
 			// Should have some size
 			assertThat(dataset.getSizeInBytes(), is(greaterThan(0L)));
 			assertThat(dataset.getSize(), is(greaterThan(0L)));
+			assertThat(getDimensions(dataFlat)[0], is(Arrays.stream(dataset.getDimensions()).reduce(1, Math::multiplyExact)));
 		} else {
 			assertThat(getDimensions(data), is(equalTo(dims)));
 			assertThat(getType(data), is(equalTo(dataset.getJavaType())));
 			// Should have some size
 			assertThat(dataset.getSizeInBytes(), is(greaterThan(0L)));
 			assertThat(dataset.getSize(), is(greaterThan(0L)));
+			assertThat(flatten(data), is(equalTo(flatten(dataFlat))));
+			assertThat(getDimensions(dataFlat).length, is(1));
+			assertThat(getDimensions(dataFlat)[0], is(Arrays.stream(dataset.getDimensions()).reduce(1, Math::multiplyExact)));
 		}
 
 		if (dataset instanceof ContiguousDataset && !dataset.isEmpty()) {