+++ jhdf/src/test/java/io/jhdf/object/datatype/FixedPointTest.java
@@ -3,7 +3,7 @@
  *
  * http://jhdf.io
  *
- * Copyright (c) 2022 James Mudd
+ * Copyright (c) 2023 James Mudd
  *
  * MIT License see 'LICENSE' file
  */
@@ -16,6 +16,7 @@ import org.junit.jupiter.api.DynamicNode;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestFactory;
 import org.junit.jupiter.api.function.Executable;
+import org.mockito.Mockito;
 
 import java.math.BigInteger;
 import java.nio.ByteBuffer;