+++ CHANGES.md
@@ -1,5 +1,20 @@
 # jHDF Change Log
 
+##  v0.6.10
+- Add support for files containing superblock extensions https://github.com/jamesmudd/jhdf/issues/462
+- Build and dependency updates
+
+## v0.6.9
+- Add support for LZ4 compressed datasets https://github.com/jamesmudd/jhdf/issues/415
+- Improve BitShuffle performance and reduce memory usage
+- Add CI on ARM64 architecture
+- Build and dependency updates
+
+## v0.6.8
+- Add support for getting flat data `Dataset#getDataFlat()` https://github.com/jamesmudd/jhdf/issues/397
+- Add support for dereferencing addresses/ids `HdfFile.getNodeByAddress(long address) ` https://github.com/jamesmudd/jhdf/issues/316
+- Dependency updates
+
 ## v0.6.7
 - Add support for Bitshuffle filter https://github.com/jamesmudd/jhdf/issues/366
 - Add ability to get filter data `Dataset#getFilters();` https://github.com/jamesmudd/jhdf/issues/378